Add a `pygmentize` binary that talks to the pygments-server.
This is a drop-in replacement for the pygments `pygmentize`.  Or at
least, drop-in from the point of view of phabricator.  This will allow
us to use phab unchanged; it will call this shell script to pygmentize
rather than the normal python script.  The hope is that dash is faster
to start up than python is, so this is a net perf win.  If not -- or
even if so -- we can always change the phab source to talk to the
server directly, it's not very hard.

Test Plan:
I ran
   make serve
and then
   ./pygmentize -O encoding=utf-8 -O stripnl=False -f html -l js < ~/khan/webapp/javascript/corelibs-package/ka-define.js
and it gave some nice-looking html as output.