Make sure a 400 error from the pygments server makes us fail.
Phabricator will notice when pygmentize fails, and will just not
pygmentize in thta case.  In our case, the typical failure is with a
400 exit code from the webserver.  But curl considers that to be a
successful curl run by default.  Adding `-f` makes it fail, which
makes the script fail, which makes phabricator ignore the output.

Auditors: benkraft

Test plan:
Ran
   ./pygmentize -l foo < /etc/issue; echo $?
and saw it say `22`