Use the default `noclasses=False`.
I don't know why I set it to true before.  Phabricator sets it to
false, and the default is false, and the pygments docs encourage you
to use False.

Auditors: benkraft

Test plan:
I deployed this manually and visited
   https://phabricator.khanacademy.org/D51932
and reloaded and selected "View options -> Highlight As -> javascript"
and saw the page render with more muted, reasonable colors.  Sadly it
didn't fix the bug with missing text, but it's a start...