Update the pygments version.
This is the most up-to-date one.  It also has an important bugfix,
apparently -- at least, after upgrading D51932 highlights properly
whereas before it would omit the "Tests for funnel manager" line on
the right hand side.  Shrug.

Auditors: benkraft

Test plan:
Installed manually on the phabricator ec2 machine, visited
   https://phabricator.khanacademy.org/D51932
and did "View options -> Highlight As -> Javascript" to force it to
re-color, and saw that the diff displayed correctly with "Tests for
FunnelManager" visible on both sides.